requirements.txt
flask
twilio
openai
python-dotenv
google-cloud-speech
pydub
gunicorn
requests